[refactor] move book utility fns to seperate file
